# Parse local cabal files concurrently