--- a/stack.cabal
+++ b/stack.cabal
@@ -176,6 +176,7 @@ library
                    , http-client-tls >= 0.3.0
                    , http-conduit >= 2.2.0
                    , http-types >= 0.8.6 && < 0.10
+                   , lifted-async
                    , lifted-base
                    , microlens >= 0.3.0.0
                    , monad-control