This should particularly benefit large multi-package projects.