@sjakobi 2016-07-08
Argh, the above commands didn't even run `getLocalPackageViews`. I had totally forgotten about my optimization in `stack clean` which will now only run `getLocalPackageViews` if the user specifies a target manually.

This makes the performance difference even more surprising though. I'll get some sleep and take another look at this.