@sjakobi 2016-07-09
I still can't see any performance gains from this, rather slight decreases. Reading all the cabal files (even when there are more than 100) concurrently might not be the best idea either.

I might look into this more once there has been a new release and we have better logs.